Update final2.py
Merged dbscan and isolation forest methods, added to main. Commented out the classify and cluster method calls for this part.
Converted DBSCAN output method to match Teddy's.
Definitely will need to clean this up more, but for the time being, I don't have the time before this is due to fix the bugs.